Fix robot not being able to vacuum with 2in1 box
In the previous logic, the robot was forced to the mop and vacuum mode
if the 2 in 1 box has been available, even though the robot did not have
the mopping attachment. This commit adds a few comments to the code and
changes the logic so that it considers whether the mopping attachment is
available or not as well as the box type.